+++ src/items/Bonemeal.java
@@ -9,9 +9,9 @@ public class Bonemeal extends Item {
 	
 	public final static String BONE_NAME = "Bonemeal";		//Display Name
 	public final static String BONE_DESC = "Spooky";		//Description
-	public final static double BONE_PRICE = 2;				//Base Price
+	public final static float BONE_PRICE = 2;				//Base Price
 	public final static String BONE_EFFECT = "growth";		//Item Effect
-	public final static double BONE_MULT = 2;
+	public final static float BONE_MULT = 2;
 	
 	public Bonemeal() {
 		super(BONE_NAME, BONE_DESC, BONE_PRICE, BONE_EFFECT, BONE_MULT);
