+++ src/items/BrewingStand.java
@@ -9,9 +9,9 @@ public class BrewingStand extends Item {
 	
 	public final static String STAND_NAME = "Brewing Stand";	
 	public final static String STAND_DESC = "Quick drink this";	
-	public final static double STAND_PRICE = 2;			
+	public final static float STAND_PRICE = 2;			
 	public final static String STAND_EFFECT = "health";		
-	public final static double STAND_MULT = 2;	
+	public final static float STAND_MULT = 2;	
 	
 	public BrewingStand() {
 		super(STAND_NAME, STAND_DESC, STAND_PRICE, STAND_EFFECT, STAND_MULT);
